Add food truck owner role

+++ app/controllers/trucks_controller.rb
@@ -10,49 +10,78 @@ class TrucksController < ApplicationController
   end
 
   def new
-    @truck = Truck.new
+    if current_user.admin? || current_user.truck_owner?
+      @truck = Truck.new
+    else
+      flash[:notice] = "You are not authorized to add a truck."
+      redirect_to trucks_path
+    end
   end
 
   def create
-    @truck = Truck.new(truck_params)
-    @truck.user = current_user
-    @truck.location = Location.find(params[:location])
+    if current_user.admin? || current_user.truck_owner?
+      @truck = Truck.new(
+        name: truck_params[:name],
+        location: Location.find(truck_params[:location_id]),
+        description: truck_params[:description]
+      )
 
-    if @truck.save
-      flash[:notice] = "Truck added successfully"
-      redirect_to truck_path(@truck)
+      @truck.user = current_user
+
+      if @truck.save
+        flash[:notice] = "Truck added successfully"
+        redirect_to truck_path(@truck)
+      else
+        flash[:notice] = "Unable to add truck"
+        render :new
+      end
     else
-      render :new
+      flash[:notice] = "You are not authorized to add a truck."
+      redirect_to trucks_path
     end
   end
 
   def edit
     @truck = Truck.find(params[:id])
+
+    if !(@truck.user == current_user || current_user.admin?)
+      flash[:notice] = "You are not authorized to update this truck's info."
+      redirect_to truck_path(@truck)
+    end
+
   end
 
   def update
     @truck = Truck.find(params[:id])
-    @truck.user = current_user
-    @truck.update(truck_params)
-    @truck.location = Location.find(params[:location])
 
-    if @truck.save
-      flash[:notice] = "Truck updated successfully"
-      redirect_to truck_path(@truck)
+    if @truck.user == current_user || current_user.admin?
+      @truck.update(truck_params)
+      @truck.location = Location.find(params[:location])
+
+      if @truck.save
+        flash[:notice] = "Truck updated successfully"
+        redirect_to truck_path(@truck)
+      else
+        render :edit
+      end
+
     else
-      render :new
+      flash[:notice] = "You are not authorized to update this truck's info."
+      redirect_to truck_path(@truck)
     end
   end
 
   def destroy
     @truck = Truck.find(params[:id])
 
-    if Truck.destroy(@truck.id)
-      Review.where(truck_id: params[:id]).destroy_all
-      flash[:notice] = "Food Truck successfully deleted"
-      redirect_to trucks_path
-    else
-      render truck_path(@truck)
+    if @truck.user == current_user || current_user.admin?
+      if Truck.destroy(@truck.id)
+        Review.where(truck_id: params[:id]).destroy_all
+        flash[:notice] = "Food Truck successfully deleted"
+        redirect_to trucks_path
+      else
+        render truck_path(@truck)
+      end
     end
   end
 
@@ -86,7 +115,7 @@ class TrucksController < ApplicationController
 
   private
   def truck_params
-    params.require(:truck).permit(:name, :description, :email, :location)
+    params.require(:truck).permit(:name, :description, :user, :location_id)
   end
 
   def find_trucks_matching_category
